+++ src/Presentation/Container/Main/index.tsx
@@ -14,10 +14,9 @@ import {
 } from "@/Presentation/Component";
 import { useRouter } from "next/router";
 import MontlyCalendar from './customCalendar.jsx';
-import { CalenderModel } from "@/Presentation/Model";
 import { SubjectType } from "@/Data/Model";
-import { CalenderEntity } from "@/Domain/Entity";
 import NewTask from "./newTask";
+import WeeklyView from "./weeklyView.jsx";
 import style from "@/Presentation/Style/Main.module.css";
 import "react-calendar/dist/Calendar.css";
 
@@ -25,12 +24,7 @@ import "react-calendar/dist/Calendar.css";
 const Main = () => {
   // const upperBarDate = showToday();
   // isTodoCheck === false, make progress bar
-  const [calender, setCalender] = useState<CalenderEntity[]>([]);
-  const [timeline, setTimeline] = useState<CalenderEntity>();
-  const [isCalenderLoading, setIsCalenderLoading] = useState(false);
-  const [isTimelineLoading, setIsTimelineLoading] = useState(true);
   const [date, setDate] = useState<Date>(new Date());
-  const cModel = new CalenderModel();
   // const router = useRouter();
   const [toggleOn, setModalOn] = useState(false);
   const [deadLine, setDeadLine] = useState("");
@@ -48,91 +42,28 @@ const Main = () => {
     setModalOn(false);
   };
 
-  // const updateNowDate = (n: number): void => {
-  //   const changeDate = date.getDay() === 0 ? 7 : date.getDay();
-  //   setDate(new Date(date.setDate(date.getDate() + (n - changeDate))));
-  // };
+  const clearDeadLine = () => {
+    setDeadLine("");
+  }
 
-  const getWeek = async () => {
-    const res = await cModel.getCalender();
-    setCalender(res);
-    setIsCalenderLoading(true);
-    await Promise.all(
-      res.map((d) => {
-        const calenderDate = new Date(
-          `${d.day.split(".")[0]}-${d.day.split(".")[1]}-${d.day.split(".")[2]
-          }`,
-        );
-        if (
-          date.getFullYear() === calenderDate.getFullYear() &&
-          date.getMonth() === calenderDate.getMonth() &&
-          date.getDate() === calenderDate.getDate()
-        )
-          setTimeline(d);
-      }),
-    );
-    setIsTimelineLoading(true);
-  };
-
-  const updateTimeLineData = () => {
-    setTimeline(undefined);
-    calender.map((d) => {
-      const calenderDate = new Date(
-        `${d.day.split(".")[0]}-${d.day.split(".")[1]}-${d.day.split(".")[2]}`,
-      );
-      if (
-        date.getFullYear() === calenderDate.getFullYear() &&
-        date.getMonth() === calenderDate.getMonth() &&
-        date.getDate() === calenderDate.getDate()
-      )
-        setTimeline(d);
-    });
-  };
 
   useEffect(() => {
-    getWeek();
-  }, []);
-
-  useEffect(() => {
-    updateTimeLineData();
-  }, [date]);
-
-  useEffect(() => {
-  },[dailyTodo]);
+  }, [dailyTodo]);
 
   return (
     <div className={style.Main}>
       <Header />
-      <div className={style.ContentBox}>
-        <div className={style.TodoDate}>
-          <p>Today : {`${date.getFullYear()}.${date.getMonth() + 1
-            }.${date.getDate()}`}</p>
-        </div>
-        {toggleOn ? (
-          <NewTask closeModal={closeModal} /> // 여기서 데드라인 보내줘야함
-        ) : typeof timeline === "undefined" ? (
-          <div>주간 일정을 로딩 중 입니다 ... </div>
-        ) : isTimelineLoading ? (
-          timeline.subject.map((todo, i) => {
-            return (
-              <Todo
-                key={i} // Adjust the key to ensure uniqueness
-                label={todo.label}
-                name={todo.name}
-                startTime={todo.startTime}
-                endTime={todo.endTime}
-                todoType={todo.label === 0 ? "fixed" : "check"}
-                prevValue={undefined}
-                checked={undefined}
-              ></Todo>
-            );
-          })
-        ) : (
-          <>
-          </>
-        )}
+      <div className={style.MonthandDay}>
+        <div className={style.ContentBox}>
+          <div className={style.TodoDate}>
+            <p>Today : {`${date.getFullYear()}.${date.getMonth() + 1
+              }.${date.getDate()}`}</p>
+          </div>
+          {toggleOn ? (
+            <NewTask closeModal={closeModal} clearDeadLine={clearDeadLine} /> // 여기서 데드라인 보내줘야함
+          ) : (<></>)}
 
-        {/* {todos.map((todoSubject, index) =>
+          {/* {todos.map((todoSubject, index) =>
           todoSubject.subject.map((todo, todoIndex) => (
             <Todo
               key={index * 10 + todoIndex} // Adjust the key to ensure uniqueness
@@ -147,24 +78,22 @@ const Main = () => {
           ))
         )} */
 
-        }
-        {
-          dailyTodo.length > 0 && <ul>
-            {dailyTodo.map((schedule:any, index)=>(
-              <li className={style.DailyList} key={index}>{schedule.name}</li>
-            ))}
-          </ul>
-        }
+          }
+          {
+            dailyTodo.length > 0 && <ul>
+              {dailyTodo.map((schedule: any, index) => (
+                <li className={style.DailyList} key={index}>{schedule.name}</li>
+              ))}
+            </ul>
+          }
+        </div>
+        <div className={style.CalenderBox}>
+            <>
+              <MontlyCalendar setDeadLine={handleDeadLine} setDailyTodo={setDailyTodo} />
+            </>
+        </div>
       </div>
-      <div className={style.CalenderBox}>
-        {isCalenderLoading ? (
-          <></>
-        ) : (
-          <>
-            <MontlyCalendar setDeadLine={handleDeadLine} setDailyTodo={setDailyTodo} />
-          </>
-        )}
-      </div>
+      <WeeklyView />
     </div>
   );
 };
