+++ src/Presentation/Style/Main.module.css
@@ -1,11 +1,23 @@
 .Main {
   width: 100vw;
-  height: 100vh;
+  background-color: #f5f5f5;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  scroll-snap-type: y mandatory;
+}
+
+.MonthandDay {
+  margin: 25vh 0;
+  width: 100vw;
   background-color: #f5f5f5;
   display: flex;
   align-items: center;
   justify-content: center;
   gap: 10px;
+  scroll-snap-align: start;
 }
 
 .ContentBox {
