+++ src/Presentation/Style/weeklyView.module.css
@@ -1,0 +1,7 @@
+.weeklyMain {
+    scroll-snap-align: start;
+    margin: 25vh 0;
+    width: 1400px;
+    height: 700px;
+    background-color: bisque;
+}
